--- src/main/java/com/work/shopping/Entity/Orders.java
@@ -1,65 +1,0 @@
-package com.work.shopping.Entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.math.BigInteger;
-import java.util.Date;
-
-@Entity
-@Table(name = "orders")
-public class Orders {
-    @Id
-    @Column(name = "id")
-    private BigInteger id;
-    @Column(name = "account")
-    private String account;
-    @Column(name = "data")
-    private String data;
-    @Column(name = "destination")
-    private String destination;
-
-    public BigInteger getId() {
-        return id;
-    }
-
-    public void setId(BigInteger id) {
-        this.id = id;
-    }
-
-    public String getAccount() {
-        return account;
-    }
-
-    public void setAccount(String account) {
-        this.account = account;
-    }
-
-    public String getData() {
-        return data;
-    }
-
-    public void setData(String data) {
-        this.data = data;
-    }
-
-    public String getDestination() {
-        return destination;
-    }
-
-    public void setDestination(String destination) {
-        this.destination = destination;
-    }
-
-    @Override
-    public String toString() {
-        return "orders{" +
-                "account=" + account +
-                ", id='" + id + '\'' +
-                ", data=" + data +
-                ", destination='" + destination + '\'' +
-                '}';
-    }
-}
-
